format short number

+++ ReservationApp.Application/Services/interfaces/IDashboardService.cs
@@ -4,7 +4,7 @@ namespace ReservationApp.Application.Services.interfaces;
 
 public interface IDashboardService
 {
-    public RadialBarChartDTO GetRadialCartDataModel(int total, double currentMonth, double prevMonth);
+    public RadialBarChartDTO GetRadialCartDataModel(double total, double currentMonth, double prevMonth);
     public RadialBarChartDTO GetRevenueRadialChartData();
     public RadialBarChartDTO GetTotalBookingRadialChartData();
     public RadialBarChartDTO GetUserRegisteredRadialChartData();
